requirements.txt
pandas==2.2.1
numpy==1.26.4
matplotlib==3.8.4
seaborn==0.13.2
scikit-learn==1.4.1
requests==2.31.0
python-dotenv==1.0.1
